Boss: Created Cutscene Phase II, created skill phase II, completed Checkpoint and upgrade power UI

+++ Assets/Script/Design Pattern/Tree Behavior/LeafNode/CheckChangePhase.cs
@@ -21,9 +21,9 @@ public class CheckChangePhase : BehaviorNode
                 return NodeState.Failure;
             }
 
-            if ((float)bossSystem.Health.currentHealth / bossSystem.Health.maxHealth <= bossSystem.UltimateCombo[bossSystem.NextPhase].HealthThreshold)
+            if (bossSystem.Health.currentHealth / bossSystem.Health.maxHealth <= bossSystem.UltimateCombo[bossSystem.NextPhase].HealthThreshold)
             {
-                return NodeState.Success;
+                    return NodeState.Success;
             }
         }
